Resolves issue #37 - Reorganize Swift files
